+++ car.hpp
@@ -6,7 +6,7 @@
 class Car{
     public:
         Car(float top_speed,
-            float lane,
+            unsigned int lane,
             float start_distance,
             float goal_distance);
         float get_top_speed();
